+++ plot.py
@@ -1,8 +1,10 @@
 """Contains functions for plotting the airfoil data."""
-
 import matplotlib.pyplot as plt
 from matplotlib.widgets import RadioButtons, Slider, TextBox
 from airfoil import *
+import asyncio
+import pandas as pd
+from tkinter import filedialog
 
 class Plotting():
     def __init__(self, df, task_list) -> None:
@@ -40,11 +42,18 @@ class Plotting():
         self.ax_x.set_visible(False)
         self.ax_y.set_visible(False)
 
+        # Add button to store results
+        self.save_button_ax = plt.axes([0.85, 0.01, 0.1, 0.03])
+        self.save_button = plt.Button(self.save_button_ax, 'Save Results')
+        self.save_button_ax.set_visible(False)
+        self.save_button.on_clicked(self.save_results)
+        self.save_results_concat = None  # Placeholder for concatenated results
+
         # Show initial plot (Task 1 by default)
         self.update_plot(task=self.task_list[0])
         plt.show()
 
-    def plot_twisted_airfoil(self, df, angle_degrees, ledge=True, centroid=False) -> None:
+    async def plot_twisted_airfoil(self, df, angle_degrees, ledge=True, centroid=False) -> None:
         """
         Plot the original and twisted airfoil
         
@@ -59,9 +68,9 @@ class Plotting():
         
         # Apply twist
         if ledge:
-            result_df = twist_airfoil_ledge(df=df, angle_degrees=angle_degrees)
+            result_df = await twist_airfoil_ledge(df=df, angle_degrees=angle_degrees)
         else:
-            result_df = twist_airfoil_centroid(df=df, angle_degrees=angle_degrees)
+            result_df = await twist_airfoil_centroid(df=df, angle_degrees=angle_degrees)
         
         # Plot original airfoil 
         plt.plot(result_df['X'], result_df['Y'], 'b-', linewidth=2, label='Original')
@@ -84,7 +93,7 @@ class Plotting():
         plt.axis('equal')
         plt.legend()
 
-    def plot_scale_airfoil(self, df, scale_factor) -> None:
+    async def plot_scale_airfoil(self, df, scale_factor) -> None:
         """
         Plot the original and scaled airfoil.
         
@@ -96,7 +105,7 @@ class Plotting():
         plt.cla()            # Clear the main axes only
         
         # Apply scale
-        result_df = scale_airfoil(df, scale_factor)
+        result_df = await scale_airfoil(df, scale_factor)
         
         # Plot original airfoil 
         plt.plot(result_df['X'], result_df['Y'], 'b-', linewidth=2, label='Original')
@@ -113,7 +122,7 @@ class Plotting():
         plt.axis('equal')
         plt.legend()
 
-    def plot_translated_airfoil(self, df, centroid_x, centroid_y):
+    async def plot_translated_airfoil(self, df, centroid_x, centroid_y):
         """
         Plot the original and translated airfoil.
         
@@ -126,14 +135,30 @@ class Plotting():
         plt.cla()            # Clear the main axes only
 
         # Apply translation
-        result_df_translated = translate_airfoil(df=df, x_center=centroid_x, y_center=centroid_y)
-        result_df_twisted = twist_airfoil_centroid(df=result_df_translated[['X_translated', 'Y_translated']], angle_degrees=self.angle_slider.val)
-        result_df_scale = scale_airfoil(df=result_df_twisted[['X_twisted', 'Y_twisted']], scale_factor=self.scale_slider.val)
+        result_df_translated = await translate_airfoil(df=df, x_center=centroid_x, y_center=centroid_y)
+        result_df_twisted = await twist_airfoil_centroid(df=result_df_translated[['X_translated', 'Y_translated']], angle_degrees=self.angle_slider.val)
+        result_df_scale = await scale_airfoil(df=result_df_twisted[['X_twisted', 'Y_twisted']], scale_factor=self.scale_slider.val)
 
+        # Store results in the DataFrame
+        self.save_results_concat = pd.concat(
+            [result_df_translated, 
+             result_df_twisted[['X_twisted', 'Y_twisted']], 
+             result_df_scale[['X_scaled', 'Y_scaled']]
+            ], axis=1)
+
+        # Plot original and transformed airfoils
         plt.plot(result_df_translated['X'], result_df_translated['Y'], 'b-', linewidth=2, label='Original')
-        plt.plot(result_df_translated['X_translated'], result_df_translated['Y_translated'], 'r--', linewidth=2)
-        plt.plot(result_df_twisted['X_twisted'], result_df_twisted['Y_twisted'], 'g-.', linewidth=2)
-        plt.plot(result_df_scale['X_scaled'], result_df_scale['Y_scaled'], 'm:', linewidth=2)
+        plt.plot(result_df_translated['X_translated'], result_df_translated['Y_translated'], 'r--', linewidth=2, label=f'Translated by ({centroid_x}, {centroid_y})')
+        plt.plot(result_df_twisted['X_twisted'], result_df_twisted['Y_twisted'], 'g-.', linewidth=2, label=f'Twisted by {self.angle_slider.val}°')
+        plt.plot(result_df_scale['X_scaled'], result_df_scale['Y_scaled'], 'm:', linewidth=2, label=f'Scaled by {self.scale_slider.val}')
+
+        # Setup plot
+        plt.grid(visible=True, linestyle='--')
+        plt.xlabel(xlabel='X', fontsize=12)
+        plt.ylabel(ylabel='Y', fontsize=12)
+        plt.title(label=f'Airfoil transformed', fontsize=14)
+        plt.axis('equal')
+        plt.legend()
 
     def update_plot(self, task) -> None:
         plt.sca(self.ax)
@@ -145,7 +170,13 @@ class Plotting():
             self.scale_slider.set_active(False)
             self.ax_x.set_visible(False)
             self.ax_y.set_visible(False)
-            self.plot_twisted_airfoil(df=self.df, angle_degrees=self.angle_slider.val, ledge=True, centroid=False)
+            asyncio.run(self.plot_twisted_airfoil(
+                df=self.df, 
+                angle_degrees=self.angle_slider.val, 
+                ledge=True, 
+                centroid=False
+                )
+            )
         elif task == 'Task 2: Scale':
             self.ax_angle.set_visible(False)
             self.angle_slider.set_active(False)
@@ -153,7 +184,11 @@ class Plotting():
             self.scale_slider.set_active(True)
             self.ax_x.set_visible(False)
             self.ax_y.set_visible(False)
-            self.plot_scale_airfoil(df=self.df, scale_factor=self.scale_slider.val)
+            asyncio.run(self.plot_scale_airfoil(
+                df=self.df, 
+                scale_factor=self.scale_slider.val
+                )
+            )
         elif task == 'Task 3: Twist by Centroid':
             self.ax_angle.set_visible(True)
             self.angle_slider.set_active(True)
@@ -161,7 +196,13 @@ class Plotting():
             self.scale_slider.set_active(False)
             self.ax_x.set_visible(False)
             self.ax_y.set_visible(False)
-            self.plot_twisted_airfoil(df=self.df, angle_degrees=self.angle_slider.val, ledge=False, centroid=True)
+            asyncio.run(self.plot_twisted_airfoil(
+                df=self.df, 
+                angle_degrees=self.angle_slider.val, 
+                ledge=False, 
+                centroid=True
+                )
+            )
         else:
             self.ax_angle.set_visible(True)
             self.angle_slider.set_active(True)
@@ -169,5 +210,21 @@ class Plotting():
             self.scale_slider.set_active(True)
             self.ax_x.set_visible(True)
             self.ax_y.set_visible(True)
-            self.plot_translated_airfoil(df=self.df, centroid_x=float(self.x_textbox.text), centroid_y=float(self.y_textbox.text))
+            self.save_button_ax.set_visible(True)
+            asyncio.run(self.plot_translated_airfoil(
+                df=self.df, 
+                centroid_x=float(self.x_textbox.text), 
+                centroid_y=float(self.y_textbox.text)
+                )
+            )
         plt.draw()
+    
+    def save_results(self, event) -> None:
+        """Save the transformed results to a file."""
+        file_path = filedialog.asksaveasfilename(
+            title='Save Transformed Airfoil Coordinates',
+            defaultextension='.txt',
+            filetypes=[('Text Files', '*.txt'), ('All Files', '*.*')]
+        )
+        if file_path:
+            self.save_results_concat.to_csv(file_path, sep='\t', index=False, float_format='%.6f')