+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG.md
 
+## 1.0.2 - 08/07/2025
+
+Added: 
+    
+    - dialog window to choose input file
+    - button and dialog window to save output file
+
+Fixed:
+
+    - task 4: twisted and scale dataframe are now able to catch the transferred dataframe
+    - refactor: create 2 seperate slider for task 4
+
+Bug to fix:
+
+    - task 4: data in save file do not align with index
+
 ## 1.0.1 - 07/07/2025
 
 Added: